northbridge_sim/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.6.1
aiosqlite==0.20.0
redis==5.2.0
duckdb==1.1.3
pyarrow==18.1.0
pandas==2.2.3
numpy==2.1.3
pyyaml==6.0.2
cryptofeed==2.4.1
alpaca-py==0.36.0
streamlit==1.41.1
streamlit-autorefresh==1.0.1
requests==2.32.3
